retraining_pipeline/requirements.txt
scikit-learn>=1.0.0
xgboost>=1.5.0
mlflow>=2.0.0
google-cloud-storage>=2.0.0
pandas>=1.3.0
numpy>=1.20.0
joblib>=1.0.0
python-dotenv>=0.19.0
requests>=2.25.0
imbalanced-learn>=0.8.0
matplotlib>=3.5.0
seaborn>=0.11.0
